commands/update: move 'is repo check' for --rev to lower level
